refactor: moved jwk and request function to new file
